vfetch: Implement in-place vertex fetch optimizer
This change implements two new functions, optimizeVertexFetchInplace and
optimizeVertexFetchInplaceMulti. They perform vertex fetch optimization
without creating a separate vertex array copy.

To do that, we need to maintain two remap arrays: we need to maintain
the information about mapping between "physical" indices (where in the
array the element is) and "logical" indices (where in the array the
element was). The mapping needs to be bidirectional since during the
update we copy information from "logical" (based on index buffer
contents) to "physical" (based on linear array scan).

The "multi" version is necessary because with that we can share memory and
remap updates between all vertex streams.

In theory this can be substantially faster for huge meshes compared to
default implementation. In practice it seems like the multi version is
noticeably slower, while the single stream version is noticeably
faster... maybe we're hitting some sort of memcpy/branch prediction
bottleneck?