vertexcodec: Add support for filter function
In order to be able to efficiently store normal/quaternion data, we need
to reconstruct Z/W after decoding.

We could do this after decodeVertexBuffer but it's inefficient to do
this since that requires a lot of extra memory bandwidth. Instead we can
reconstruct the data in L1 cache right after it's been transposed.

This change adds support for filter callback and two basic functions to
reconstruct Z/W; the functions will need SIMD implementations.