vertexcodec: Fix filter callback application
The code for filters themselves was functioning correctly, but filters
damaged the last_vertex that was being copied out after the fact, which
led to incorrect decoding of vertices that follow after.

Also add an Emscripten function pointer helper since I'm not sure how to
use this from JS otherwise.